Line up coordinates of Psi/dMdt terms w/ G terms
Because G terms are computed as the difference of integrated tendency terms, they lie on the (target grid) interfaces rather than the levels that the tendency terms themselves lie on. Similarly, because the Psi and dMdt terms are integrals below an interface, they lie also lie on the interfaces while the cell thicknesses (binned into the target tracer coordinate grid) lie on the level surfaces. Previously, I was computing these two different kinds of terms on staggered output grids which made their direct comparison challenging. I have now modified the discretization approach (and fixed coordinate labeling errors) to make the two computations self-consistent.

Additionally, I now use `xgcm.grid.transform` with `target_data` on the diagnostic tracer cell coordinates, as opposed to first interpolating the target data onto vertical coordinate interfaces. Because the transformations are conservative, what matters is just that the `target` grid is appropriately defined. It is actually much cleaner and more robust to use `target_data` on the same grid as the `data` that is being binned!