Exactly reproduces key `xwmb` results but with G terms from `xwmt`
`calc_wmt_theta` was renamed to `calc_wmt_theta_old` and replaced with a new
`calc_wmt_theta` function that requires importing `xwmt.wmt` (specifically,
commit https://github.com/hdrake/xwmt/commit/1ba193a7245a32f523098ffb97f0e2a98c676590
was used here).

The notebook computes the full water mass budget (in temperature space only, for now) and shows
that the new algorithm exactly reproduces the old results! Yay! The old algorithm will be deprecated
in one of the next commits; `xwmb` will instead use `xwmt` as its only method of calculating the
water mass transformation terms going forward.

Renamed a bunch of the diagnostic variables to be consistent with PEP8 conventions (all lower
case variable and function names) and to be a bit more descrptive.

Was getting some NaNs corrupting calculations when merging all of the different variables
into a common dataset because the target grid `thetao_i` only agrees to machine precision
(~1.e-16) and not *exactly*, which is necessary for a perfect merge. I added a check that
the target coordinates are close (to within rtol=1.e-14) which then overrides them to be
exactly the same, or else gives a ValueError.