Change budgets_dict input to full budget, from simple budgets used by xwmt
This is necessary because `xwmb` needs specific terms in the budget, which can conflict with
those needed by `xwmt` for the water mass transformations. For example, one might want to decompose
the surface_exchange_flux terms, but `xwmb` needs the sum of the surface_exchange_flux for mass to
compute the `mass_source` term in the budget. Similar issue for the process_material_derivative needed
for numerical mixing!

Other changes:
- Option for assert_zero_transport for computations where the transport is known to be zero,
either to save computation time or because (umo,vmo) diagnostics are not available.