Demonstrate problem with MOM6 budgets
Difference in mass/heat/salt contents between two timesteps is inconsistent
with volume-integrated tendencies, but they should be exactly the same.

This was a notebook I used to check whether I could reproduce these errors.
I convinced myself that the errors was not that big for the purposes of
large-scale water mass transformation budgets.

Keith Lindsay's related PR:
https://github.com/NCAR/MOM6/pull/215