Switch default `dt` units to seconds from ns
The variable `dt = ds.time_bounds.diff("time_bounds")` is used
to compute the finite difference dM/dt (`wmt.mass_tendency`).
This was assumed implicit assumed to be in units of nanoseconds, because
for MOM6 data the time-difference has dtype `timedelta64[ns]`.

Instead, the default now assumed `dt` to have units of seconds and only
converts from nanoseconds to seconds if the dtype is `'<m8[ns]'`.